Move helper method to concerns
